Please be confident of New Test for V2
I just saw the new version of the code, and tested several lengths of input Strings, I think this version is correct, both two time complexities are O(n) and both two space complexities are O(1), this outcome can make sense. I'll finish the excel table this moring, please be confident,and we need to start the slides as soon as possible!